Use fslock to synchronize multiprocess log rolling

+++ src/append/rolling_file/policy/compound/roll/fixed_window.rs
@@ -89,6 +89,7 @@ pub struct FixedWindowRoller {
     count: u32,
     #[cfg(feature = "background_rotation")]
     cond_pair: Arc<(Mutex<bool>, Condvar)>,
+    lockfile: PathBuf,
 }
 
 impl FixedWindowRoller {
@@ -101,6 +102,11 @@ impl FixedWindowRoller {
 impl Roll for FixedWindowRoller {
     #[cfg(not(feature = "background_rotation"))]
     fn roll(&self, file: &Path) -> anyhow::Result<()> {
+        let mut lockfile = fslock::LockFile::open(&self.lockfile)?;
+        if !lockfile.try_lock()? {
+            return Ok(()) // another process is already rolling
+        }
+
         if self.count == 0 {
             return fs::remove_file(file).map_err(Into::into);
         }
@@ -282,6 +288,7 @@ impl FixedWindowRollerBuilder {
             count,
             #[cfg(feature = "background_rotation")]
             cond_pair: Arc::new((Mutex::new(true), Condvar::new())),
+            lockfile: PathBuf::from(pattern).with_extension("lock"),
         })
     }
 }
